feat: exporing OidcData (needed in auth-server)
