+++ packages/sdk/src/client/oidc/index.ts
@@ -1,2 +1,3 @@
 export * from "./account.js";
 export * from "./client.js";
+export * from "./serialize.js";
