+++ packages/sdk/src/client/oidc/serialize.ts
@@ -1,0 +1,9 @@
+import type { Hex } from "viem";
+
+export type OidcData = {
+  oidcDigest: Hex;
+  iss: string;
+  readyToRecover: boolean;
+  pendingPasskeyHash: Hex;
+  recoverNonce: Hex;
+};
